Moved RefEq to quasar.qscript.analysis
