+++ connector/src/main/scala/quasar/qscript/QScriptCore.scala
@@ -21,6 +21,7 @@ import quasar.{NonTerminal, Terminal, RenderTree, RenderTreeT}, RenderTree.ops._
 import quasar.common.SortDir
 import quasar.contrib.matryoshka._
 import quasar.fp._
+import quasar.qscript.analysis.RefEq
 
 import matryoshka._
 import matryoshka.data._
